% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.R
\name{model}
\alias{model}
\title{ODE solution}
\usage{
model(t, y, param)
}
\arguments{
\item{user}{created parameters}
}
\value{
solution of an ODE function system
}
\description{
This function solves an ODE system based on the given parameter list.
}
