% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.R
\name{model_and_plot}
\alias{model_and_plot}
\title{SEIR}
\usage{
model_and_plot()
}
\value{
Returns a plot of the SEIR model simulation.
}
\description{
This function runs an SEIR model simulation and plots the results.
}
